+++ packages/stream_chat_flutter/lib/src/dialogs/channel_info_dialog.dart
@@ -39,7 +39,7 @@ class ChannelInfoDialog extends StatelessWidget {
             children: [
               StreamChannelInfo(
                 channel: channel,
-                textStyle: StreamChatTheme.of(context).channelPreviewTheme.subtitleStyle,
+                textStyle: context.streamTextTheme.captionDefault,
               ),
             ],
           ),