+++ packages/stream_chat_flutter/lib/src/theme/themes.dart
@@ -1,7 +1,6 @@
 export 'avatar_theme.dart';
 export 'channel_header_theme.dart';
 export 'channel_list_header_theme.dart';
-export 'channel_preview_theme.dart';
 export 'color_theme.dart';
 export 'draft_list_tile_theme.dart';
 export 'gallery_footer_theme.dart';